+++ src/API/View.js
@@ -35,6 +35,7 @@ class View {
     }
     // is the pointer stale?
     if (this.frozen && isSame(docs, this.frozen) === false) {
+      console.log('=== repair docs ==')
       repair(this)
       docs = world.methods.one.getDoc(this.pointer, this.document)
     }
@@ -71,6 +72,7 @@ class View {
   get fullPointer() {
     let { docs, pointer, document, frozen } = this
     if (frozen && isSame(docs, this.frozen) === false) {
+      console.log('=== repair pointer ==')
       repair(this)
       docs = this.docs
     }