+++ src/API/methods/freeze.js
@@ -1,8 +1,12 @@
 const methods = {
   // allow re-use of this view, after a mutation
   freeze: function () {
-    // this.compute('index')
     this.frozen = this.docs
+    return this
+  },
+  // make it fast again
+  unfreeze: function () {
+    this.frozen = null
     return this
   },
 }